Replaced Date with Timestamp Replaced float with BigDecimal Replaced int with long

+++ plugins/eCommerce/src/main/java/com/skytala/eCommerce/entity/Product.java
@@ -1,5 +1,7 @@
 package com.skytala.eCommerce.entity;
 
+
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Map;
 
@@ -32,26 +34,26 @@ private String inventoryMessage;
 private String inventoryItemTypeId;
 private Boolean requireInventory;
 private String quantityUomId;
-private Float quantityIncluded;
-private Integer piecesIncluded;
+private BigDecimal quantityIncluded;
+private Long piecesIncluded;
 private Boolean requireAmount;
-private Float fixedAmount;
+private BigDecimal fixedAmount;
 private String amountUomTypeId;
 private String weightUomId;
-private Float shippingWeight;
-private Float productWeight;
+private BigDecimal shippingWeight;
+private BigDecimal productWeight;
 private String heightUomId;
-private Float productHeight;
-private Float shippingHeight;
+private BigDecimal productHeight;
+private BigDecimal shippingHeight;
 private String widthUomId;
-private Float productWidth;
-private Float shippingWidth;
+private BigDecimal productWidth;
+private BigDecimal shippingWidth;
 private String depthUomId;
-private Float productDepth;
-private Float shippingDepth;
+private BigDecimal productDepth;
+private BigDecimal shippingDepth;
 private String diameterUomId;
-private Float productDiameter;
-private Float productRating;
+private BigDecimal productDiameter;
+private BigDecimal productRating;
 private String ratingTypeEnum;
 private Boolean returnable;
 private Boolean taxable;
@@ -63,10 +65,10 @@ private Boolean isVariant;
 private String virtualVariantMethodEnum;
 private String originGeoId;
 private String requirementMethodEnumId;
-private Integer billOfMaterialLevel;
-private Float reservMaxPersons;
-private Float reserv2ndPPPerc;
-private Float reservNthPPPerc;
+private Long billOfMaterialLevel;
+private BigDecimal reservMaxPersons;
+private BigDecimal reserv2ndPPPerc;
+private BigDecimal reservNthPPPerc;
 private String configId;
 private Date createdDate;
 private String createdByUserLogin;
@@ -78,9 +80,15 @@ private String lotIdFilledIn;
 private Boolean orderDecimalQuantity;
 
 
+
+
+
 	public String getProductId() {
 	return productId;
 }
+
+
+
 
 
 public void setProductId(String productId) {
@@ -88,9 +96,15 @@ public void setProductId(String productId) {
 }
 
 
+
+
+
 public String getProductTypeId() {
 	return productTypeId;
 }
+
+
+
 
 
 public void setProductTypeId(String productTypeId) {
@@ -98,9 +112,15 @@ public void setProductTypeId(String productTypeId) {
 }
 
 
+
+
+
 public String getPrimaryProductCategoryId() {
 	return primaryProductCategoryId;
 }
+
+
+
 
 
 public void setPrimaryProductCategoryId(String primaryProductCategoryId) {
@@ -108,9 +128,15 @@ public void setPrimaryProductCategoryId(String primaryProductCategoryId) {
 }
 
 
+
+
+
 public String getManufacturerPartyId() {
 	return manufacturerPartyId;
 }
+
+
+
 
 
 public void setManufacturerPartyId(String manufacturerPartyId) {
@@ -118,9 +144,15 @@ public void setManufacturerPartyId(String manufacturerPartyId) {
 }
 
 
+
+
+
 public String getFacilityId() {
 	return facilityId;
 }
+
+
+
 
 
 public void setFacilityId(String facilityId) {
@@ -128,9 +160,15 @@ public void setFacilityId(String facilityId) {
 }
 
 
+
+
+
 public Date getIntroductionDate() {
 	return introductionDate;
 }
+
+
+
 
 
 public void setIntroductionDate(Date introductionDate) {
@@ -138,9 +176,15 @@ public void setIntroductionDate(Date introductionDate) {
 }
 
 
+
+
+
 public Date getReleaseDate() {
 	return releaseDate;
 }
+
+
+
 
 
 public void setReleaseDate(Date releaseDate) {
@@ -148,9 +192,15 @@ public void setReleaseDate(Date releaseDate) {
 }
 
 
+
+
+
 public Date getSupportDiscontinuationDate() {
 	return supportDiscontinuationDate;
 }
+
+
+
 
 
 public void setSupportDiscontinuationDate(Date supportDiscontinuationDate) {
@@ -158,9 +208,15 @@ public void setSupportDiscontinuationDate(Date supportDiscontinuationDate) {
 }
 
 
+
+
+
 public Date getSalesDiscontinuationDate() {
 	return salesDiscontinuationDate;
 }
+
+
+
 
 
 public void setSalesDiscontinuationDate(Date salesDiscontinuationDate) {
@@ -168,9 +224,15 @@ public void setSalesDiscontinuationDate(Date salesDiscontinuationDate) {
 }
 
 
+
+
+
 public Boolean getSalesDiscWhenNotAvail() {
 	return salesDiscWhenNotAvail;
 }
+
+
+
 
 
 public void setSalesDiscWhenNotAvail(Boolean salesDiscWhenNotAvail) {
@@ -178,9 +240,15 @@ public void setSalesDiscWhenNotAvail(Boolean salesDiscWhenNotAvail) {
 }
 
 
+
+
+
 public String getInternalName() {
 	return internalName;
 }
+
+
+
 
 
 public void setInternalName(String internalName) {
@@ -188,9 +256,15 @@ public void setInternalName(String internalName) {
 }
 
 
+
+
+
 public String getBrandName() {
 	return brandName;
 }
+
+
+
 
 
 public void setBrandName(String brandName) {
@@ -198,9 +272,15 @@ public void setBrandName(String brandName) {
 }
 
 
+
+
+
 public String getComments() {
 	return comments;
 }
+
+
+
 
 
 public void setComments(String comments) {
@@ -208,9 +288,15 @@ public void setComments(String comments) {
 }
 
 
+
+
+
 public String getProductName() {
 	return productName;
 }
+
+
+
 
 
 public void setProductName(String productName) {
@@ -218,9 +304,15 @@ public void setProductName(String productName) {
 }
 
 
+
+
+
 public String getDescription() {
 	return description;
 }
+
+
+
 
 
 public void setDescription(String description) {
@@ -228,9 +320,15 @@ public void setDescription(String description) {
 }
 
 
+
+
+
 public String getLongDescription() {
 	return longDescription;
 }
+
+
+
 
 
 public void setLongDescription(String longDescription) {
@@ -238,9 +336,15 @@ public void setLongDescription(String longDescription) {
 }
 
 
+
+
+
 public String getPriceDetailText() {
 	return priceDetailText;
 }
+
+
+
 
 
 public void setPriceDetailText(String priceDetailText) {
@@ -248,9 +352,15 @@ public void setPriceDetailText(String priceDetailText) {
 }
 
 
+
+
+
 public String getSmallImageUrl() {
 	return smallImageUrl;
 }
+
+
+
 
 
 public void setSmallImageUrl(String smallImageUrl) {
@@ -258,9 +368,15 @@ public void setSmallImageUrl(String smallImageUrl) {
 }
 
 
+
+
+
 public String getMediumImageUrl() {
 	return mediumImageUrl;
 }
+
+
+
 
 
 public void setMediumImageUrl(String mediumImageUrl) {
@@ -268,9 +384,15 @@ public void setMediumImageUrl(String mediumImageUrl) {
 }
 
 
+
+
+
 public String getLargeImageUrl() {
 	return largeImageUrl;
 }
+
+
+
 
 
 public void setLargeImageUrl(String largeImageUrl) {
@@ -278,9 +400,15 @@ public void setLargeImageUrl(String largeImageUrl) {
 }
 
 
+
+
+
 public String getDetailImageUrl() {
 	return detailImageUrl;
 }
+
+
+
 
 
 public void setDetailImageUrl(String detailImageUrl) {
@@ -288,9 +416,15 @@ public void setDetailImageUrl(String detailImageUrl) {
 }
 
 
+
+
+
 public String getOriginalImageUrl() {
 	return originalImageUrl;
 }
+
+
+
 
 
 public void setOriginalImageUrl(String originalImageUrl) {
@@ -298,9 +432,15 @@ public void setOriginalImageUrl(String originalImageUrl) {
 }
 
 
+
+
+
 public String getDetailScreen() {
 	return detailScreen;
 }
+
+
+
 
 
 public void setDetailScreen(String detailScreen) {
@@ -308,9 +448,15 @@ public void setDetailScreen(String detailScreen) {
 }
 
 
+
+
+
 public String getInventoryMessage() {
 	return inventoryMessage;
 }
+
+
+
 
 
 public void setInventoryMessage(String inventoryMessage) {
@@ -318,9 +464,15 @@ public void setInventoryMessage(String inventoryMessage) {
 }
 
 
+
+
+
 public String getInventoryItemTypeId() {
 	return inventoryItemTypeId;
 }
+
+
+
 
 
 public void setInventoryItemTypeId(String inventoryItemTypeId) {
@@ -328,9 +480,15 @@ public void setInventoryItemTypeId(String inventoryItemTypeId) {
 }
 
 
+
+
+
 public Boolean getRequireInventory() {
 	return requireInventory;
 }
+
+
+
 
 
 public void setRequireInventory(Boolean requireInventory) {
@@ -338,9 +496,15 @@ public void setRequireInventory(Boolean requireInventory) {
 }
 
 
+
+
+
 public String getQuantityUomId() {
 	return quantityUomId;
 }
+
+
+
 
 
 public void setQuantityUomId(String quantityUomId) {
@@ -348,24 +512,39 @@ public void setQuantityUomId(String quantityUomId) {
 }
 
 
-public Float getQuantityIncluded() {
+
+
+
+public BigDecimal getQuantityIncluded() {
 	return quantityIncluded;
 }
 
 
-public void setQuantityIncluded(Float quantityIncluded) {
+
+
+
+public void setQuantityIncluded(BigDecimal quantityIncluded) {
 	this.quantityIncluded = quantityIncluded;
 }
 
 
-public Integer getPiecesIncluded() {
+
+
+
+public Long getPiecesIncluded() {
 	return piecesIncluded;
 }
 
 
-public void setPiecesIncluded(Integer piecesIncluded) {
+
+
+
+public void setPiecesIncluded(Long piecesIncluded) {
 	this.piecesIncluded = piecesIncluded;
 }
+
+
+
 
 
 public Boolean getRequireAmount() {
@@ -373,19 +552,31 @@ public Boolean getRequireAmount() {
 }
 
 
+
+
+
 public void setRequireAmount(Boolean requireAmount) {
 	this.requireAmount = requireAmount;
 }
 
 
-public Float getFixedAmount() {
+
+
+
+public BigDecimal getFixedAmount() {
 	return fixedAmount;
 }
 
 
-public void setFixedAmount(Float fixedAmount) {
+
+
+
+public void setFixedAmount(BigDecimal fixedAmount) {
 	this.fixedAmount = fixedAmount;
 }
+
+
+
 
 
 public String getAmountUomTypeId() {
@@ -393,9 +584,15 @@ public String getAmountUomTypeId() {
 }
 
 
+
+
+
 public void setAmountUomTypeId(String amountUomTypeId) {
 	this.amountUomTypeId = amountUomTypeId;
 }
+
+
+
 
 
 public String getWeightUomId() {
@@ -403,29 +600,47 @@ public String getWeightUomId() {
 }
 
 
+
+
+
 public void setWeightUomId(String weightUomId) {
 	this.weightUomId = weightUomId;
 }
 
 
-public Float getShippingWeight() {
+
+
+
+public BigDecimal getShippingWeight() {
 	return shippingWeight;
 }
 
 
-public void setShippingWeight(Float shippingWeight) {
+
+
+
+public void setShippingWeight(BigDecimal shippingWeight) {
 	this.shippingWeight = shippingWeight;
 }
 
 
-public Float getProductWeight() {
+
+
+
+public BigDecimal getProductWeight() {
 	return productWeight;
 }
 
 
-public void setProductWeight(Float productWeight) {
+
+
+
+public void setProductWeight(BigDecimal productWeight) {
 	this.productWeight = productWeight;
 }
+
+
+
 
 
 public String getHeightUomId() {
@@ -433,29 +648,47 @@ public String getHeightUomId() {
 }
 
 
+
+
+
 public void setHeightUomId(String heightUomId) {
 	this.heightUomId = heightUomId;
 }
 
 
-public Float getProductHeight() {
+
+
+
+public BigDecimal getProductHeight() {
 	return productHeight;
 }
 
 
-public void setProductHeight(Float productHeight) {
+
+
+
+public void setProductHeight(BigDecimal productHeight) {
 	this.productHeight = productHeight;
 }
 
 
-public Float getShippingHeight() {
+
+
+
+public BigDecimal getShippingHeight() {
 	return shippingHeight;
 }
 
 
-public void setShippingHeight(Float shippingHeight) {
+
+
+
+public void setShippingHeight(BigDecimal shippingHeight) {
 	this.shippingHeight = shippingHeight;
 }
+
+
+
 
 
 public String getWidthUomId() {
@@ -463,29 +696,47 @@ public String getWidthUomId() {
 }
 
 
+
+
+
 public void setWidthUomId(String widthUomId) {
 	this.widthUomId = widthUomId;
 }
 
 
-public Float getProductWidth() {
+
+
+
+public BigDecimal getProductWidth() {
 	return productWidth;
 }
 
 
-public void setProductWidth(Float productWidth) {
+
+
+
+public void setProductWidth(BigDecimal productWidth) {
 	this.productWidth = productWidth;
 }
 
 
-public Float getShippingWidth() {
+
+
+
+public BigDecimal getShippingWidth() {
 	return shippingWidth;
 }
 
 
-public void setShippingWidth(Float shippingWidth) {
+
+
+
+public void setShippingWidth(BigDecimal shippingWidth) {
 	this.shippingWidth = shippingWidth;
 }
+
+
+
 
 
 public String getDepthUomId() {
@@ -493,29 +744,47 @@ public String getDepthUomId() {
 }
 
 
+
+
+
 public void setDepthUomId(String depthUomId) {
 	this.depthUomId = depthUomId;
 }
 
 
-public Float getProductDepth() {
+
+
+
+public BigDecimal getProductDepth() {
 	return productDepth;
 }
 
 
-public void setProductDepth(Float productDepth) {
+
+
+
+public void setProductDepth(BigDecimal productDepth) {
 	this.productDepth = productDepth;
 }
 
 
-public Float getShippingDepth() {
+
+
+
+public BigDecimal getShippingDepth() {
 	return shippingDepth;
 }
 
 
-public void setShippingDepth(Float shippingDepth) {
+
+
+
+public void setShippingDepth(BigDecimal shippingDepth) {
 	this.shippingDepth = shippingDepth;
 }
+
+
+
 
 
 public String getDiameterUomId() {
@@ -523,29 +792,47 @@ public String getDiameterUomId() {
 }
 
 
+
+
+
 public void setDiameterUomId(String diameterUomId) {
 	this.diameterUomId = diameterUomId;
 }
 
 
-public Float getProductDiameter() {
+
+
+
+public BigDecimal getProductDiameter() {
 	return productDiameter;
 }
 
 
-public void setProductDiameter(Float productDiameter) {
+
+
+
+public void setProductDiameter(BigDecimal productDiameter) {
 	this.productDiameter = productDiameter;
 }
 
 
-public Float getProductRating() {
+
+
+
+public BigDecimal getProductRating() {
 	return productRating;
 }
 
 
-public void setProductRating(Float productRating) {
+
+
+
+public void setProductRating(BigDecimal productRating) {
 	this.productRating = productRating;
 }
+
+
+
 
 
 public String getRatingTypeEnum() {
@@ -553,14 +840,23 @@ public String getRatingTypeEnum() {
 }
 
 
+
+
+
 public void setRatingTypeEnum(String ratingTypeEnum) {
 	this.ratingTypeEnum = ratingTypeEnum;
 }
 
 
-	public Boolean getReturnable() {
+
+
+
+public Boolean getReturnable() {
 	return returnable;
 }
+
+
+
 
 
 public void setReturnable(Boolean returnable) {
@@ -568,9 +864,15 @@ public void setReturnable(Boolean returnable) {
 }
 
 
+
+
+
 public Boolean getTaxable() {
 	return taxable;
 }
+
+
+
 
 
 public void setTaxable(Boolean taxable) {
@@ -578,9 +880,15 @@ public void setTaxable(Boolean taxable) {
 }
 
 
+
+
+
 public Boolean getChargeShipping() {
 	return chargeShipping;
 }
+
+
+
 
 
 public void setChargeShipping(Boolean chargeShipping) {
@@ -588,9 +896,15 @@ public void setChargeShipping(Boolean chargeShipping) {
 }
 
 
+
+
+
 public Boolean getAutoCreateKeywords() {
 	return autoCreateKeywords;
 }
+
+
+
 
 
 public void setAutoCreateKeywords(Boolean autoCreateKeywords) {
@@ -598,9 +912,15 @@ public void setAutoCreateKeywords(Boolean autoCreateKeywords) {
 }
 
 
+
+
+
 public Boolean getIncludeInPromotions() {
 	return includeInPromotions;
 }
+
+
+
 
 
 public void setIncludeInPromotions(Boolean includeInPromotions) {
@@ -608,9 +928,15 @@ public void setIncludeInPromotions(Boolean includeInPromotions) {
 }
 
 
+
+
+
 public Boolean getIsVirtual() {
 	return isVirtual;
 }
+
+
+
 
 
 public void setIsVirtual(Boolean isVirtual) {
@@ -618,9 +944,15 @@ public void setIsVirtual(Boolean isVirtual) {
 }
 
 
+
+
+
 public Boolean getIsVariant() {
 	return isVariant;
 }
+
+
+
 
 
 public void setIsVariant(Boolean isVariant) {
@@ -628,9 +960,15 @@ public void setIsVariant(Boolean isVariant) {
 }
 
 
+
+
+
 public String getVirtualVariantMethodEnum() {
 	return virtualVariantMethodEnum;
 }
+
+
+
 
 
 public void setVirtualVariantMethodEnum(String virtualVariantMethodEnum) {
@@ -638,9 +976,15 @@ public void setVirtualVariantMethodEnum(String virtualVariantMethodEnum) {
 }
 
 
+
+
+
 public String getOriginGeoId() {
 	return originGeoId;
 }
+
+
+
 
 
 public void setOriginGeoId(String originGeoId) {
@@ -648,9 +992,15 @@ public void setOriginGeoId(String originGeoId) {
 }
 
 
+
+
+
 public String getRequirementMethodEnumId() {
 	return requirementMethodEnumId;
 }
+
+
+
 
 
 public void setRequirementMethodEnumId(String requirementMethodEnumId) {
@@ -658,44 +1008,71 @@ public void setRequirementMethodEnumId(String requirementMethodEnumId) {
 }
 
 
-public Integer getBillOfMaterialLevel() {
+
+
+
+public Long getBillOfMaterialLevel() {
 	return billOfMaterialLevel;
 }
 
 
-public void setBillOfMaterialLevel(Integer billOfMaterialLevel) {
+
+
+
+public void setBillOfMaterialLevel(Long billOfMaterialLevel) {
 	this.billOfMaterialLevel = billOfMaterialLevel;
 }
 
 
-public Float getReservMaxPersons() {
+
+
+
+public BigDecimal getReservMaxPersons() {
 	return reservMaxPersons;
 }
 
 
-public void setReservMaxPersons(Float reservMaxPersons) {
+
+
+
+public void setReservMaxPersons(BigDecimal reservMaxPersons) {
 	this.reservMaxPersons = reservMaxPersons;
 }
 
 
-public Float getReserv2ndPPPerc() {
+
+
+
+public BigDecimal getReserv2ndPPPerc() {
 	return reserv2ndPPPerc;
 }
 
 
-public void setReserv2ndPPPerc(Float reserv2ndPPPerc) {
+
+
+
+public void setReserv2ndPPPerc(BigDecimal reserv2ndPPPerc) {
 	this.reserv2ndPPPerc = reserv2ndPPPerc;
 }
 
 
-public Float getReservNthPPPerc() {
+
+
+
+public BigDecimal getReservNthPPPerc() {
 	return reservNthPPPerc;
 }
 
 
-public void setReservNthPPPerc(Float reservNthPPPerc) {
+
+
+
+public void setReservNthPPPerc(BigDecimal reservNthPPPerc) {
 	this.reservNthPPPerc = reservNthPPPerc;
 }
+
+
+
 
 
 public String getConfigId() {
@@ -703,9 +1080,15 @@ public String getConfigId() {
 }
 
 
+
+
+
 public void setConfigId(String configId) {
 	this.configId = configId;
 }
+
+
+
 
 
 public Date getCreatedDate() {
@@ -713,9 +1096,15 @@ public Date getCreatedDate() {
 }
 
 
+
+
+
 public void setCreatedDate(Date createdDate) {
 	this.createdDate = createdDate;
 }
+
+
+
 
 
 public String getCreatedByUserLogin() {
@@ -723,9 +1112,15 @@ public String getCreatedByUserLogin() {
 }
 
 
+
+
+
 public void setCreatedByUserLogin(String createdByUserLogin) {
 	this.createdByUserLogin = createdByUserLogin;
 }
+
+
+
 
 
 public Date getLastModifiedDate() {
@@ -733,9 +1128,15 @@ public Date getLastModifiedDate() {
 }
 
 
+
+
+
 public void setLastModifiedDate(Date lastModifiedDate) {
 	this.lastModifiedDate = lastModifiedDate;
 }
+
+
+
 
 
 public String getLastModifiedByUserLogin() {
@@ -743,9 +1144,15 @@ public String getLastModifiedByUserLogin() {
 }
 
 
+
+
+
 public void setLastModifiedByUserLogin(String lastModifiedByUserLogin) {
 	this.lastModifiedByUserLogin = lastModifiedByUserLogin;
 }
+
+
+
 
 
 public Boolean getInShippingBox() {
@@ -753,9 +1160,15 @@ public Boolean getInShippingBox() {
 }
 
 
+
+
+
 public void setInShippingBox(Boolean inShippingBox) {
 	this.inShippingBox = inShippingBox;
 }
+
+
+
 
 
 public String getDefaultShipmentBoxTypeId() {
@@ -763,9 +1176,15 @@ public String getDefaultShipmentBoxTypeId() {
 }
 
 
+
+
+
 public void setDefaultShipmentBoxTypeId(String defaultShipmentBoxTypeId) {
 	this.defaultShipmentBoxTypeId = defaultShipmentBoxTypeId;
 }
+
+
+
 
 
 public String getLotIdFilledIn() {
@@ -773,9 +1192,15 @@ public String getLotIdFilledIn() {
 }
 
 
+
+
+
 public void setLotIdFilledIn(String lotIdFilledIn) {
 	this.lotIdFilledIn = lotIdFilledIn;
 }
+
+
+
 
 
 public Boolean getOrderDecimalQuantity() {
@@ -783,17 +1208,19 @@ public Boolean getOrderDecimalQuantity() {
 }
 
 
+
+
+
 public void setOrderDecimalQuantity(Boolean orderDecimalQuantity) {
 	this.orderDecimalQuantity = orderDecimalQuantity;
 }
+
+
+
 
 
 	public Map<String, Object> mapAttributeField(){
 
 		return ProductMapper.map(this);
 	}
-
-
-
-
 }
